Expose types. Fix `saveToFs`
