+++ src/__spec__/jsonRawReport.mock.ts
@@ -1,4 +1,4 @@
-import { IJsonRawReport } from 'porchmark';
+import { IJsonRawReport } from '../types/porchmark';
 
 export default <IJsonRawReport>({
     "sites": [
